refactor(codegen): extract std::io::udp::* into stdlib/io_udp.rs (Round 1.14)
20 fns lifted into IoUdpStdlib<'ctx>: bind, send, recv, recv_with_source,
recv_into, set_recv_timeout, set_send_timeout (set_timeout family),
last_source_host, last_source_port, join_group / leave_group,
set_multicast_ttl, set_multicast_loop, set_multicast_iface,
set_option_int, set_option_bool, get_option_int, close + the three
internal udp_group_fallible / udp_set_int_fallible / udp_setsockopt_fallible
helpers used by the multicast and set_option lowerings.

codegen.rs: 40906 -> 39702 (-1204 lines).
